# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Vizsgaremek Szerver API Dokumentáció
  description: A Vizsgaremek végpontjainek részeletes leírása és használata
servers:
  - url: http://localhost:3115
paths:
  /user:
    get:
      description: Returns a list of users
      security:
        - bearerAuth: []      
      responses:
        "200":
          description: List of users returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: Creates a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: User has been created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/search:
    get:
      description: Returns the list of searched users
      security:
        - bearerAuth: []      
      parameters: 
        - name: username
          in: query
          schema:
            type: string
          example: 'ahuske2'
        - name: last_name
          in: query
          schema:
            type: string
          example: 'Huske'
        - name: first_name
          in: query
          schema:
            type: string            
          example: 'Ameline'
        - name: email
          in: query
          schema:
            type: string
          example: 'ahuske2@uiuc.edu'
        - name: role
          in: query
          schema:
            type: number
          example: 5
      responses:
        "200":
          description: List of users returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/{_id}:
    get:
      description: Get one user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          example: "62afe4aab6513601bc8660f4"
          schema:
            type: string
      responses:
        "200":
          description: Returns a user
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/User"
        "404":
          description: Requested user doesn't exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      description: Updates a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User has been updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: Removes a user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: _id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User has been deleted successfully
          content:
            application/json:
              schema:
                type: object
                example: {}
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:  
       - username
       - last_name      
       - first_name      
       - email
       - role
      properties:
        username:
          type: string
        last_name:
          type: string
        first_name:
          type: string
        email:
          type: string
        role:
          type: number
          default: 1        
...
    Error:
      type: object
      required:
        - message
        - has_error
      properties:
        message:
          type: string
        has_error:
          type: boolean